Staunch some warnings including SequenceManager.cpp
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13279 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/sound/RingBuffer.h
@@ -291,7 +291,7 @@ RingBuffer<T, N>::resized(int newSize, int R) const
     while (r != w) {
         T value = m_buffer[r];
         newBuffer->write(&value, 1);
-        if (++r == m_size) r = 0;
+        if (++r == (int)m_size) r = 0;
     }
 
     return newBuffer;
